backend/requirements.txt
fastapi
uvicorn
numpy
pillow
torch
torchvision
faiss-cpu
